% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{bmi}
\alias{bmi}
\title{Calculate Body Mass Index}
\usage{
bmi(
  weight.kg = NULL,
  height.m = NULL,
  bmi = NULL,
  categ = FALSE,
  labels = FALSE
)
}
\arguments{
\item{weight.kg}{weight in kg}

\item{height.m}{height in m}

\item{bmi}{body mass index already computed}

\item{categ}{logical indicating whether to return a factor
(who classification)}

\item{labels}{if \code{categ == TRUE}, logical indicating whether to
return factor with labels (who classification)}
}
\value{
a vector of body mass index
}
\description{
Calculate Body Mass Index
}
\examples{

## Simple calculation
bmi(w = 85, h = 1.85)
## Who classes without or with labels
bmi(b = 24, categ = TRUE)
bmi(b = 27, categ = TRUE, labels = TRUE)

}
