% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sf12}
\alias{sf12}
\title{SF12 questionnaire scoring}
\usage{
sf12(X = NULL)
}
\arguments{
\item{X}{a \code{\link{matrix}} or \code{\link{data.frame}} of 12
columns, containing questionnaire items. In order from left to right: gh1, pf02, pf04, rp2,
rp3, re2, re3, bp2, mh3, vt2, mh4, sf2.}
}
\description{
SF12 questionnaire scoring
}
\note{
This is an R port of SAS algorithm by Apolone and Mosconi found
\href{http://crc.marionegri.it/qdv/index.php?page=sf12}{here}.

SF-12  is a registered trademark of medical outcomes trust.
}
\examples{
scores <- sf12(sf12sample)
## website data test (printing with many decimals for 10 selected
## questionnaires)
web <- c(1,2,4,5,11,27,28,31,37,39)
print(scores[web,], digits = 6)
## SF12 Manual checks
print(unlist(lapply(scores, mean)), digits = 3)
print(unlist(lapply(scores, sd)), digits = 3)
print(lapply(scores, range), digits = 3)
## Correlations
db <- cbind(sf12sample, scores)
var.order <- c(2:5,8,1,10,12,6,7,9,11)
cors <- cor(db)[var.order, 13:14]
print(cors, digits = 1)
## Fine: reversed item have reverse sign correlation coefficients
}

