% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf12.R
\docType{data}
\name{sf12sample}
\alias{sf12sample}
\title{SF12 sample dataset}
\format{
A data frame with 50 rows and 13 columns (1 id columns, the
other SF12 items)
}
\source{
SF12 italian manual at
\url{http://crc.marionegri.it/qdv/index.php?page=sf12}
}
\usage{
sf12sample
}
\description{
SF12 sample of 50 questionnaires \code{sf12} testing purposes.
}
\keyword{datasets}
