% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sf36}
\alias{sf36}
\title{SF-36 questionnaire scoring}
\usage{
sf36(X = NULL)
}
\arguments{
\item{X}{a \code{\link{matrix}} or \code{\link{data.frame}} of 12
columns, containing questionnaire items. In order from left to right: gh1, pf02, pf04, rp2,
rp3, re2, re3, bp2, mh3, vt2, mh4, sf2.}
}
\description{
SF-36 questionnaire scoring
}
\note{
This is an R port of SAS algorithm by Apolone and Mosconi found
\href{http://crc.marionegri.it/qdv/index.php?page=sf36}{here}.

SF-12  is a registered trademark of medical outcomes trust.
}
\examples{
\dontrun{
## -------------------------
## Algorithm test/validation
## -------------------------
}
scores <- sf36(sf36sample)
}

