% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{sf36sample}
\alias{sf36sample}
\title{SF-36 sample dataset}
\format{A data frame with 10 rows and 36 columns (1 per
SF-36 items)}
\source{
SF-36 italian SAS algorithm available at
\url{http://crc.marionegri.it/qdv/index.php?page=sf36}
}
\usage{
sf36sample
}
\description{
SF-36 sample of 10 questionnaires for \code{\link{sf36}} testing purposes.
}
\keyword{datasets}

