% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stai_Y1}
\alias{stai_Y1}
\title{State-trait Anxiety inventory for Adults Scoreing Key (Form Y1)}
\usage{
stai_Y1(Q = NULL)
}
\arguments{
\item{Q}{a matrix or data.frame of 20 columns (corresponing in a ordered
way to each, so therefore Q[,1] is the first item of the scale) filled
by 1 (Not at all) to 4 (Very much so) values}
}
\description{
State-trait Anxiety inventory for Adults Scoreing Key (Form Y1)
}

