% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stai_Y2}
\alias{stai_Y2}
\title{State-trait Anxiety inventory for Adults Scoreing Key (Form Y2)}
\usage{
stai_Y2(Q = NULL)
}
\arguments{
\item{Q}{a matrix or data.frame of 20 columns (corresponing in a ordered
way to each, so therefore Q[,1] is the first item of the scale) filled
by 1 (Almost never) to 4 (Almost always) values}
}
\description{
State-trait Anxiety inventory for Adults Scoreing Key (Form Y2)
}

