% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf12.R
\name{sf12}
\alias{sf12}
\title{SF12 questionnaire scoring}
\usage{
sf12(X = NULL)
}
\arguments{
\item{X}{a \code{\link{matrix}} or \code{\link{data.frame}} of 12
columns, containing questionnaire items. In order from left to right:
GH1 (In generale direbbe che la Sua salute è),
PF02 (Attività di moderato impegno fisico),                                                                       
PF04 (Salire qualche piano di scale),                                                                             
RP2 (Ha reso meno di quanto avrebbe voluto),                                                                     
RP3 (Ha dovuto limitare alcuni tipi di lavoro),                                                                  
RE2 (Ha reso meno di quanto avrebbe voluto),                                                                     
RE3 (Ha avuto un calo di concentrazione sul lavoro),                                                             
BP2 (In che misura il dolore la ha ostacolata nel lavoro che svolge abitualmente),                               
MH3 (Calmo e sereno),                                                                                            
VT2 (Pieno di energia),                                                                                          
MH4 (Scoraggiato e triste),                                                                                      
SF2 (Per quanto tempo la Sua salute fisica o il Suo stato emotivo hanno interferito nelle Sue attività sociali)}
}
\description{
SF12 questionnaire scoring
}
\note{
This is an R port of SAS algorithm by Apolone and Mosconi found
\href{https://lsi.marionegri.it/qdv/index.php?page=sf12}{here}.

SF-12  is a registered trademark of medical outcomes trust.
}
\examples{
## -------------------------
## Algorithm test/validation
## -------------------------
(scores <- sf12(sf12sample))
## website data test (printing with many decimals for 10 selected
## questionnaires)
web <- c(1,2,4,5,11,27,28,31,37,39)
print(scores[web,], digits = 6)
## SF12 Manual checks
print(unlist(lapply(scores, mean)), digits = 3)
print(unlist(lapply(scores, sd)), digits = 3)
print(lapply(scores, range), digits = 3)
## Correlations
db <- cbind(sf12sample, scores)
var.order <- c(2:5,8,1,10,12,6,7,9,11)
cors <- cor(db)[var.order, 13:14]
print(cors, digits = 1)
## Fine: reversed item have reverse sign correlation coefficients 
}
